Check if node with label name "label" exists

+++ platform-controller/src/main/java/org/hobbit/controller/kubernetes/K8sClusterManagerImpl.java
@@ -78,7 +78,7 @@ public class K8sClusterManagerImpl implements K8sClusterManager {
     }
 
     @Override
-    public long getNumberOfNodes(String label) {
+    public int getNumberOfNodes(String label) {
 
         Lister<V1Node> nodeLister = new Lister<V1Node>(nodeInformer.getIndexer());
         V1Node node = nodeLister.get(label);
